Refactoring the new code into multiple new files
