+++ src/AzurePipelinesToGitHubActionsConverter.Core/Conversion/GeneralProcessing.cs
@@ -1,6 +1,5 @@
 ﻿using AzurePipelinesToGitHubActionsConverter.Core.AzurePipelines;
 using AzurePipelinesToGitHubActionsConverter.Core.Conversion.Serialization;
-using AzurePipelinesToGitHubActionsConverter.Core.GitHubActions;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
@@ -11,12 +10,10 @@ namespace AzurePipelinesToGitHubActionsConverter.Core.Conversion
 {
     public class GeneralProcessing
     {
-        public List<string> VariableList;
         public string MatrixVariableName;
         private readonly bool _verbose;
         public GeneralProcessing(bool verbose)
         {
-            VariableList = new List<string>();
             _verbose = verbose;
         }
 
@@ -25,46 +22,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Core.Conversion
             if (nameYaml != null)
             {
                 return nameYaml.Replace("name:", "").Replace(System.Environment.NewLine, "").Trim();
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        public GitHubActions.Trigger ProcessTriggerV2(string triggerYaml)
-        {
-            AzurePipelines.Trigger trigger = null;
-            if (triggerYaml != null)
-            {
-                try
-                {
-                    string[] simpleTrigger = GenericObjectSerialization.DeserializeYaml<string[]>(triggerYaml);
-                    trigger = new AzurePipelines.Trigger
-                    {
-                        branches = new IncludeExclude
-                        {
-                            include = simpleTrigger
-                        }
-                    };
-                }
-                catch (Exception ex)
-                {
-                    Debug.WriteLine($"DeserializeYaml<string[]>(triggerYaml) swallowed an exception: " + ex.Message);
-                    trigger = GenericObjectSerialization.DeserializeYaml<AzurePipelines.Trigger>(triggerYaml);
-                }
-            }
-
-            //Convert the pieces to GitHub
-            GitHubActions.Trigger push = ProcessComplexTrigger(trigger);
-
-            //Build the return results
-            if (push != null)
-            {
-                return new GitHubActions.Trigger
-                {
-                    push = push?.push
-                };
             }
             else
             {
@@ -85,7 +42,7 @@ namespace AzurePipelinesToGitHubActionsConverter.Core.Conversion
                 }
                 catch (Exception ex)
                 {
-                    Debug.WriteLine($"DeserializeYaml<string[]>(triggerYaml) swallowed an exception: " + ex.Message);
+                    Debug.WriteLine($"DeserializeYaml<string>(dependsOnYaml) swallowed an exception: " + ex.Message);
                     dependsOn = GenericObjectSerialization.DeserializeYaml<string[]>(dependsOnYaml);
                 }
             }
@@ -163,346 +120,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Core.Conversion
             return environment;
         }
 
-        public GitHubActions.Trigger ProcessPullRequestV2(string pullRequestYaml)
-        {
-            AzurePipelines.Trigger trigger = null;
-            if (pullRequestYaml != null)
-            {
-                try
-                {
-                    string[] simpleTrigger = GenericObjectSerialization.DeserializeYaml<string[]>(pullRequestYaml);
-                    trigger = new AzurePipelines.Trigger
-                    {
-                        branches = new IncludeExclude
-                        {
-                            include = simpleTrigger
-                        }
-                    };
-                }
-                catch (Exception ex)
-                {
-                    Debug.WriteLine($"DeserializeYaml<string[]>(pullRequestYaml) swallowed an exception: " + ex.Message);
-                    trigger = GenericObjectSerialization.DeserializeYaml<AzurePipelines.Trigger>(pullRequestYaml);
-                }
-            }
-
-            //Convert the pieces to GitHub
-            GitHubActions.Trigger pr = ProcessPullRequest(trigger);
-
-            //Build the return results
-            if (pr != null)
-            {
-                return new GitHubActions.Trigger
-                {
-                    pull_request = pr?.pull_request
-                };
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        public GitHubActions.Trigger ProcessSchedulesV2(string schedulesYaml)
-        {
-            Schedule[] schedules = null;
-            if (schedulesYaml != null)
-            {
-                try
-                {
-                    schedules = GenericObjectSerialization.DeserializeYaml<Schedule[]>(schedulesYaml);
-                }
-                catch (Exception ex)
-                {
-                    Debug.WriteLine($"DeserializeYaml<Schedule[]>(schedulesYaml) swallowed an exception: " + ex.Message);
-                }
-            }
-
-            //Convert the pieces to GitHub
-            string[] schedule = ProcessSchedules(schedules);
-
-            //Build the return results
-            if (schedule != null)
-            {
-                return new GitHubActions.Trigger
-                {
-                    schedule = schedule
-                };
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        public Dictionary<string, string> ProcessParametersAndVariablesV2(string parametersYaml, string variablesYaml)
-        {
-            List<Variable> parameters = null;
-            if (parametersYaml != null)
-            {
-                try
-                {
-                    Dictionary<string, string> simpleParameters = GenericObjectSerialization.DeserializeYaml<Dictionary<string, string>>(parametersYaml);
-                    parameters = new List<AzurePipelines.Variable>();
-                    foreach (KeyValuePair<string, string> item in simpleParameters)
-                    {
-                        parameters.Add(new Variable
-                        {
-                            name = item.Key,
-                            value = item.Value
-                        });
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Debug.WriteLine($"DeserializeYaml<Dictionary<string, string>>(parametersYaml) swallowed an exception: " + ex.Message);
-                    parameters = GenericObjectSerialization.DeserializeYaml<List<Variable>>(parametersYaml);
-                }
-            }
-
-            List<Variable> variables = null;
-            if (variablesYaml != null)
-            {
-                try
-                {
-                    Dictionary<string, string> simpleVariables = GenericObjectSerialization.DeserializeYaml<Dictionary<string, string>>(variablesYaml);
-                    variables = new List<AzurePipelines.Variable>();
-                    foreach (KeyValuePair<string, string> item in simpleVariables)
-                    {
-                        variables.Add(new Variable
-                        {
-                            name = item.Key,
-                            value = item.Value
-                        });
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Debug.WriteLine($"DeserializeYaml<Dictionary<string, string>>(variablesYaml) swallowed an exception: " + ex.Message);
-                    parameters = GenericObjectSerialization.DeserializeYaml<List<Variable>>(variablesYaml);
-                }
-            }
-
-
-            Dictionary<string, string> env = new Dictionary<string, string>();
-            Dictionary<string, string> processedParameters = ProcessComplexVariablesV2(parameters);
-            Dictionary<string, string> processedVariables = ProcessComplexVariablesV2(variables);
-            foreach (KeyValuePair<string, string> item in processedParameters)
-            {
-                if (env.ContainsKey(item.Key) == false)
-                {
-                    env.Add(item.Key, item.Value);
-                }
-            }
-            foreach (KeyValuePair<string, string> item in processedVariables)
-            {
-                if (env.ContainsKey(item.Key) == false)
-                {
-                    env.Add(item.Key, item.Value);
-                }
-            }
-
-            if (env.Count > 0)
-            {
-                return env;
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        public Dictionary<string, GitHubActions.Job> ProcessStagesV2(JToken stagesJson, string strategyYaml)
-        {
-            AzurePipelines.Job[] jobs = null;
-            List<AzurePipelines.Stage> stages = new List<AzurePipelines.Stage>();
-            if (stagesJson != null)
-            {
-                //for each stage
-                foreach (JToken stageJson in stagesJson)
-                {
-                    AzurePipelines.Stage stage = new AzurePipelines.Stage();
-                    if (stageJson["stage"] != null)
-                    {
-                        stage.stage = stageJson["stage"].ToString();
-                    }
-                    if (stageJson["displayName"] != null)
-                    {
-                        stage.displayName = stageJson["displayName"].ToString();
-                    }
-                    if (stageJson["dependsOn"] != null)
-                    {
-                        stage.dependsOn = ProcessDependsOnV2(stageJson["dependsOn"].ToString());
-                    }
-                    if (stageJson["condition"] != null)
-                    {
-                        stage.condition = stageJson["condition"].ToString();
-                    }
-                    if (stageJson["variables"] != null)
-                    {
-                        stage.variables = ProcessParametersAndVariablesV2(null, stageJson["variables"].ToString());
-                    }
-                    if (stageJson["jobs"] != null)
-                    {
-                        stage.jobs = ExtractAzurePipelinesJobsV2(stageJson["jobs"], strategyYaml);
-                    }
-                    stages.Add(stage);
-                }
-
-                //process the jobs
-                if (stages != null)
-                {
-                    int jobCount = 0;
-                    foreach (Stage stage in stages)
-                    {
-                        if (stage.jobs != null)
-                        {
-                            jobCount += stage.jobs.Length;
-                        }
-                    }
-                    jobs = new AzurePipelines.Job[jobCount];
-
-                    int jobIndex = 0;
-                    foreach (Stage stage in stages)
-                    {
-                        if (stage.jobs != null)
-                        {
-                            for (int i = 0; i < stage.jobs.Length; i++)
-                            {
-                                jobs[jobIndex] = stage.jobs[i];
-                                if (stage.variables != null)
-                                {
-                                    if (jobs[jobIndex].variables == null)
-                                    {
-                                        jobs[jobIndex].variables = new Dictionary<string, string>();
-                                    }
-                                    foreach (KeyValuePair<string, string> stageVariable in stage.variables)
-                                    {
-                                        //Add the stage variable if it doesn't already exist
-                                        if (jobs[jobIndex].variables.ContainsKey(stageVariable.Key) == false)
-                                        {
-                                            jobs[jobIndex].variables.Add(stageVariable.Key, stageVariable.Value);
-                                        }
-                                    }
-                                }
-                                if (stage.condition != null)
-                                {
-                                    jobs[jobIndex].condition = stage.condition;
-                                }
-                                //Get the job name
-                                string jobName = ConversionUtility.GenerateJobName(stage.jobs[i], jobIndex);
-                                //Rename the job, using the stage name as prefix, so that we keep the job names unique
-                                jobs[jobIndex].job = stage.stage + "_Stage_" + jobName;
-                                jobIndex++;
-                            }
-                        }
-                    }
-                }
-            }
-
-            if (jobs != null)
-            {
-                Dictionary<string, GitHubActions.Job> gitHubJobs = new Dictionary<string, GitHubActions.Job>();
-                foreach (AzurePipelines.Job job in jobs)
-                {
-                    JobProcessing jobProcessing = new JobProcessing(_verbose);
-                    gitHubJobs.Add(job.job, jobProcessing.ProcessJob(job, null));
-                }
-                return gitHubJobs;
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        public AzurePipelines.Job[] ExtractAzurePipelinesJobsV2(JToken jobsJson, string strategyYaml)
-        {
-            AzurePipelines.Job[] jobs = new AzurePipelines.Job[jobsJson.Count()];
-            if (jobsJson != null)
-            {
-                int i = 0;
-                foreach (JToken jobJson in jobsJson)
-                {
-                    AzurePipelines.Job job = new AzurePipelines.Job();
-                    if (jobJson["job"] != null)
-                    {
-                        job.job = jobJson["job"].ToString();
-                    }
-                    if (jobJson["deployment"] != null)
-                    {
-                        job.deployment = jobJson["deployment"].ToString();
-                    }
-                    if (jobJson["displayName"] != null)
-                    {
-                        job.displayName = jobJson["displayName"].ToString();
-                    }
-                    if (jobJson["pool"] != null)
-                    {
-                        job.pool = ProcessPoolV2(jobJson["pool"].ToString());
-                    }
-                    if (jobJson["strategy"] != null)
-                    {
-                        job.strategy = ProcessStrategyV2(jobJson["strategy"].ToString());
-                    }
-                    else if (strategyYaml != null)
-                    {
-                        job.strategy = ProcessStrategyV2(strategyYaml);
-                    }
-                    if (jobJson["dependsOn"] != null)
-                    {
-                        job.dependsOn = ProcessDependsOnV2(jobJson["dependsOn"].ToString());
-                    }
-                    if (jobJson["condition"] != null)
-                    {
-                        job.condition = ConditionsProcessing.TranslateConditions(jobJson["condition"].ToString());
-                    }
-                    if (jobJson["template"] != null)
-                    {
-                        job.template = jobJson["template"].ToString();
-                    }
-                    if (jobJson["environment"] != null)
-                    {
-                        job.environment = ProcessEnvironmentV2(jobJson["environment"].ToString());
-                    }
-                    if (jobJson["timeoutInMinutes"] != null)
-                    {
-                        int.TryParse(jobJson["timeoutInMinutes"].ToString(), out int timeOut);
-                        if (timeOut > 0)
-                        {
-                            job.timeoutInMinutes = timeOut;
-                        }
-                    }
-                    if (jobJson["continueOnError"] != null)
-                    {
-                        bool.TryParse(jobJson["continueOnError"].ToString(), out bool continueOnError);
-                        job.continueOnError = continueOnError;
-                    }
-                    if (jobJson["variables"] != null)
-                    {
-                        job.variables = ProcessParametersAndVariablesV2(null, jobJson["variables"].ToString());
-                    }
-                    if (jobJson["steps"] != null)
-                    {
-                        //job.steps = ProcessStepsV2(jobJson["steps"]);
-                        try
-                        {
-                            job.steps = GenericObjectSerialization.DeserializeYaml<AzurePipelines.Step[]>(jobJson["steps"].ToString());
-                        }
-                        catch (Exception ex)
-                        {
-                            Debug.WriteLine($"DeserializeYaml<AzurePipelines.Step[]>(jobJson[\"steps\"].ToString() swallowed an exception: " + ex.Message);
-                            //jobs = GenericObjectSerialization.DeserializeYaml<Dictionary<string, GitHubActions.Job>>(jobYaml);
-                        }
-                    }
-                    jobs[i] = job;
-                    i++;
-                }
-            }
-
-            return jobs;
-        }
-
         public Resources ExtractResourcesV2(string resourcesYaml)
         {
             if (resourcesYaml != null)
@@ -537,85 +154,24 @@ namespace AzurePipelinesToGitHubActionsConverter.Core.Conversion
             }
             return null;
         }
+   
 
-        public Dictionary<string, GitHubActions.Job> ProcessJobsV2(AzurePipelines.Job[] jobs, Resources resources)
+        //process the build pool/agent
+        public string ProcessPool(Pool pool)
         {
-            if (jobs != null)
+            string newPool = null;
+            if (pool != null)
             {
-                Dictionary<string, GitHubActions.Job> gitHubJobs = new Dictionary<string, GitHubActions.Job>();
-                int i = 0;
-                foreach (AzurePipelines.Job job in jobs)
+                if (pool.vmImage != null)
                 {
-                    JobProcessing jobProcessing = new JobProcessing(_verbose);
-                    string jobName = job.job;
-                    if (jobName == null && job.deployment != null)
-                    {
-                        jobName = job.deployment;
-                    }
-                    else if (jobName == null && job.template != null)
-                    {
-                        jobName = "job_" + (i + 1).ToString() + "_template";
-                    }
-                    gitHubJobs.Add(jobName, jobProcessing.ProcessJob(job, resources));
-                    MatrixVariableName = jobProcessing.MatrixVariableName;
-                    i++;
+                    newPool = pool.vmImage;
                 }
-                return gitHubJobs;
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        public AzurePipelines.Job[] ProcessJobFromPipelineRootV2(string poolYaml, string strategyYaml, string stepsYaml)
-        {
-            Pool pool = null;
-            if (poolYaml != null)
-            {
-                pool = ProcessPoolV2(poolYaml);
-            }
-            AzurePipelines.Strategy strategy = null;
-            try
-            {
-                //Most often, the pool will be in this structure
-                strategy = GenericObjectSerialization.DeserializeYaml<AzurePipelines.Strategy>(strategyYaml);
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"DeserializeYaml<AzurePipelines.Strategy>(strategyYaml) swallowed an exception: " + ex.Message);
-            }
-            AzurePipelines.Step[] steps = null;
-            if (stepsYaml != null)
-            {
-                try
+                else if (pool.name != null)
                 {
-                    steps = GenericObjectSerialization.DeserializeYaml<AzurePipelines.Step[]>(stepsYaml);
-                }
-                catch (Exception ex)
-                {
-                    Debug.WriteLine($"DeserializeYaml<AzurePipelines.Step[]>(stepsYaml) swallowed an exception: " + ex.Message);
+                    newPool = pool.name;
                 }
             }
-
-            AzurePipelines.Job job = new AzurePipelines.Job
-            {
-                pool = pool,
-                strategy = strategy,
-                steps = steps
-            };
-            //Don't add the build name unless there is content
-            if (job.pool != null || job.strategy != null || steps != null)
-            {
-                AzurePipelines.Job[] jobs = new AzurePipelines.Job[1];
-                job.job = "build";
-                jobs[0] = job;
-                return jobs;
-            }
-            else
-            {
-                return null;
-            }
+            return newPool;
         }
 
         public Pool ProcessPoolV2(string poolYaml)
@@ -679,157 +235,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Core.Conversion
             return pool;
         }
 
-        //Process a simple trigger, e.g. "Trigger: [master, develop]"
-        public GitHubActions.Trigger ProcessSimpleTrigger(string[] trigger)
-        {
-            AzurePipelines.Trigger newTrigger = new AzurePipelines.Trigger
-            {
-                branches = new IncludeExclude
-                {
-                    include = trigger
-                }
-            };
-            return ProcessComplexTrigger(newTrigger);
-        }
-
-        //Process a complex trigger, using the Trigger object
-        public GitHubActions.Trigger ProcessComplexTrigger(AzurePipelines.Trigger trigger)
-        {
-            if (trigger == null)
-            {
-                return null;
-            }
-            //Note: as of 18-Oct, you receive an error if you try to post both a "branches" and a "ignore-branches", or a "paths and a ignore-paths". You can only have one or the other...
-            TriggerDetail push = new TriggerDetail();
-            //process branches
-            if (trigger.branches != null)
-            {
-                if (trigger.branches.include != null)
-                {
-                    push.branches = trigger.branches.include;
-                }
-                else if (trigger.branches.exclude != null)
-                {
-                    push.branches_ignore = trigger.branches.exclude;
-                }
-            }
-            //process paths
-            if (trigger.paths != null)
-            {
-                if (trigger.paths.include != null)
-                {
-                    push.paths = trigger.paths.include;
-                }
-                if (trigger.paths.exclude != null)
-                {
-                    push.paths_ignore = trigger.paths.exclude;
-                }
-            }
-            //process tags
-            if (trigger.tags != null)
-            {
-                if (trigger.tags.include != null)
-                {
-                    push.tags = trigger.tags.include;
-                }
-                if (trigger.tags.exclude != null)
-                {
-                    push.tags_ignore = trigger.tags.exclude;
-                }
-            }
-
-            return new GitHubActions.Trigger
-            {
-                push = push
-            };
-
-        }
-
-        //process the pull request
-        public GitHubActions.Trigger ProcessPullRequest(AzurePipelines.Trigger pr)
-        {
-            if (pr == null)
-            {
-                return null;
-            }
-            TriggerDetail pullRequest = new TriggerDetail();
-            //process branches
-            if (pr.branches != null)
-            {
-                if (pr.branches.include != null)
-                {
-                    pullRequest.branches = pr.branches.include;
-                }
-                else if (pr.branches.exclude != null)
-                {
-                    pullRequest.branches_ignore = pr.branches.exclude;
-                }
-            }
-            //process paths
-            if (pr.paths != null)
-            {
-                if (pr.paths.include != null)
-                {
-                    pullRequest.paths = pr.paths.include;
-                }
-                if (pr.paths.exclude != null)
-                {
-                    pullRequest.paths_ignore = pr.paths.exclude;
-                }
-            }
-            //process tags
-            if (pr.tags != null)
-            {
-                if (pr.tags.include != null)
-                {
-                    pullRequest.tags = pr.tags.include;
-                }
-                if (pr.tags.exclude != null)
-                {
-                    pullRequest.tags_ignore = pr.tags.exclude;
-                }
-            }
-
-            return new GitHubActions.Trigger
-            {
-                pull_request = pullRequest
-            };
-        }
-
-        //process the schedule
-        public string[] ProcessSchedules(AzurePipelines.Schedule[] schedules)
-        {
-            if (schedules == null)
-            {
-                return null;
-            }
-            string[] newSchedules = new string[schedules.Length];
-            for (int i = 0; i < schedules.Length; i++)
-            {
-                newSchedules[i] = "cron: '" + schedules[i].cron + "'";
-            }
-
-            return newSchedules;
-        }
-
-        //process the build pool/agent
-        public string ProcessPool(Pool pool)
-        {
-            string newPool = null;
-            if (pool != null)
-            {
-                if (pool.vmImage != null)
-                {
-                    newPool = pool.vmImage;
-                }
-                else if (pool.name != null)
-                {
-                    newPool = pool.name;
-                }
-            }
-            return newPool;
-        }
-
         //process the strategy matrix
         public GitHubActions.Strategy ProcessStrategy(AzurePipelines.Strategy strategy)
         {
@@ -866,7 +271,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Core.Conversion
                     string[] matrix = new string[strategy.matrix.Count];
                     KeyValuePair<string, Dictionary<string, string>> matrixVariable = strategy.matrix.First();
                     MatrixVariableName = matrixVariable.Value.Keys.First();
-                    //VariableList.Add("$(" + _matrixVariableName + ")");
                     int i = 0;
                     foreach (KeyValuePair<string, Dictionary<string, string>> entry in strategy.matrix)
                     {
@@ -961,252 +365,7 @@ namespace AzurePipelinesToGitHubActionsConverter.Core.Conversion
                 return null;
             }
         }
-
-        //process all (simple) variables
-        public Dictionary<string, string> ProcessSimpleVariables(Dictionary<string, string> variables)
-        {
-            if (variables != null)
-            {
-                //update variables from the $(variableName) format to ${{variableName}} format, by piping them into a list for replacement later.
-                foreach (string item in variables.Keys)
-                {
-                    VariableList.Add(item);
-                }
-            }
-
-            return variables;
-        }
-
-        public Dictionary<string, string> ProcessComplexVariablesV2(List<AzurePipelines.Variable> variables)
-        {
-            Dictionary<string, string> processedVariables = new Dictionary<string, string>();
-            if (variables != null)
-            {
-                //update variables from the $(variableName) format to ${{variableName}} format, by piping them into a list for replacement later.
-                for (int i = 0; i < variables.Count; i++)
-                {
-                    //name/value pairs
-                    if (variables[i].name != null && variables[i].value != null)
-                    {
-                        processedVariables.Add(variables[i].name, variables[i].value);
-                        VariableList.Add(variables[i].name);
-                    }
-                    //groups
-                    if (variables[i].group != null)
-                    {
-                        if (!processedVariables.ContainsKey("group"))
-                        {
-                            processedVariables.Add("group", variables[i].group);
-                        }
-                        else
-                        {
-                            ConversionUtility.WriteLine("group: only 1 variable group is supported at present", _verbose);
-                        }
-                    }
-                    //template
-                    if (variables[i].template != null)
-                    {
-                        processedVariables.Add("template", variables[i].template);
-                    }
-                }
-
-            }
-            return processedVariables;
-        }
-
-        //process all (complex) variables
-        public Dictionary<string, string> ProcessComplexVariables(AzurePipelines.Variable[] variables)
-        {
-            Dictionary<string, string> processedVariables = new Dictionary<string, string>();
-            if (variables != null)
-            {
-                //update variables from the $(variableName) format to ${{variableName}} format, by piping them into a list for replacement later.
-                for (int i = 0; i < variables.Length; i++)
-                {
-                    //name/value pairs
-                    if (variables[i].name != null && variables[i].value != null)
-                    {
-                        processedVariables.Add(variables[i].name, variables[i].value);
-                        VariableList.Add(variables[i].name);
-                    }
-                    //groups
-                    if (variables[i].group != null)
-                    {
-                        if (!processedVariables.ContainsKey("group"))
-                        {
-                            processedVariables.Add("group", variables[i].group);
-                        }
-                        else
-                        {
-                            ConversionUtility.WriteLine("group: only 1 variable group is supported at present", _verbose);
-                        }
-                    }
-                    //template
-                    if (variables[i].template != null)
-                    {
-                        processedVariables.Add("template", variables[i].template);
-                    }
-                }
-
-            }
-            return processedVariables;
-        }
-
-        public List<string> SearchForVariablesV2(GitHubActionsRoot gitHubActions)
-        {
-            List<string> variables = new List<string>();
-            if (gitHubActions.env != null)
-            {
-                foreach (KeyValuePair<string, string> env in gitHubActions.env)
-                {
-                    variables.Add(env.Key);
-                }
-            }
-            if (gitHubActions.jobs != null)
-            {
-                foreach (KeyValuePair<string, GitHubActions.Job> job in gitHubActions.jobs)
-                {
-                    if (job.Value.env != null)
-                    {
-                        foreach (KeyValuePair<string, string> env in job.Value.env)
-                        {
-                            variables.Add(env.Key);
-                        }
-                    }
-                }
-            }
-            return variables;
-        }
-
-
-        //process the steps
-        public GitHubActions.Step[] ProcessSteps(AzurePipelines.Step[] steps, bool addCheckoutStep = true)
-        {
-            StepsProcessing stepsProcessing = new StepsProcessing();
-
-            GitHubActions.Step[] newSteps = null;
-            if (steps != null)
-            {
-                //Start by scanning all of the steps, to see if we need to insert additional tasks
-                int stepAdjustment = 0;
-                bool addJavaSetupStep = false;
-                bool addGradleSetupStep = false;
-                bool addAzureLoginStep = false;
-                bool addMSSetupStep = false;
-                string javaVersion = null;
-
-                //If the code needs a Checkout step, add it first
-                if (addCheckoutStep == true)
-                {
-                    stepAdjustment++; // we are inserting a step and need to start moving steps 1 place into the array
-                }
-
-                //Loop through the steps to see if we need other tasks inserted in for specific circumstances
-                foreach (AzurePipelines.Step step in steps)
-                {
-                    if (step.task != null)
-                    {
-                        switch (step.task.ToUpper()) //Set to upper case to handle case sensitivity comparisons e.g. NPM hangles Npm, NPM, or npm. 
-                        {
-                            //If we have an Java based step, we will need to add a Java setup step
-                            case "ANT@1":
-                            case "MAVEN@3":
-                                if (addJavaSetupStep == false)
-                                {
-                                    addJavaSetupStep = true;
-                                    stepAdjustment++;
-                                    javaVersion = stepsProcessing.GetStepInput(step, "jdkVersionOption");
-                                }
-                                break;
-
-                            //Needs a the Java step and an additional Gradle step
-                            case "GRADLE@2":
-
-                                if (addJavaSetupStep == false)
-                                {
-                                    addJavaSetupStep = true;
-                                    stepAdjustment++;
-                                    //Create the java step, as it doesn't exist
-                                    javaVersion = "1.8";
-                                }
-                                if (addGradleSetupStep == false)
-                                {
-                                    addGradleSetupStep = true;
-                                    stepAdjustment++;
-                                }
-                                break;
-
-                            //If we have an Azure step, we will need to add a Azure login step
-                            case "AZUREAPPSERVICEMANAGE@0":
-                            case "AZURERESOURCEGROUPDEPLOYMENT@2":
-                            case "AZURERMWEBAPPDEPLOYMENT@3":
-                                if (addAzureLoginStep == false)
-                                {
-                                    addAzureLoginStep = true;
-                                    stepAdjustment++;
-                                }
-                                break;
-
-                            case "VSBUILD@1":
-                                if (addMSSetupStep == false)
-                                {
-                                    addMSSetupStep = true;
-                                    stepAdjustment++;
-                                }
-                                break;
-                        }
-                    }
-                }
-
-                //Re-size the newSteps array with adjustments as needed
-                newSteps = new GitHubActions.Step[steps.Length + stepAdjustment];
-
-                int adjustmentsUsed = 0;
-
-                //Add the steps array
-                if (addCheckoutStep == true)
-                {
-                    //Add the check out step to get the code
-                    newSteps[adjustmentsUsed] = stepsProcessing.CreateCheckoutStep();
-                    adjustmentsUsed++;
-                }
-                if (addJavaSetupStep == true)
-                {
-                    //Add the JavaSetup step to the code
-                    if (javaVersion != null)
-                    {
-                        newSteps[adjustmentsUsed] = stepsProcessing.CreateSetupJavaStep(javaVersion);
-                        adjustmentsUsed++;
-                    }
-                }
-                if (addGradleSetupStep == true)
-                {
-                    //Add the Gradle setup step to the code
-                    newSteps[adjustmentsUsed] = stepsProcessing.CreateSetupGradleStep();
-                    adjustmentsUsed++;
-                }
-                if (addAzureLoginStep == true)
-                {
-                    //Add the Azure login step to the code
-                    newSteps[adjustmentsUsed] = stepsProcessing.CreateAzureLoginStep();
-                    adjustmentsUsed++;
-                }
-                if (addMSSetupStep == true)
-                {
-                    //Add the Azure login step to the code
-                    newSteps[adjustmentsUsed] = stepsProcessing.CreateMSBuildSetupStep();
-                    //adjustmentsUsed++;
-                }
-
-                //Translate the other steps
-                for (int i = stepAdjustment; i < steps.Length + stepAdjustment; i++)
-                {
-                    newSteps[i] = stepsProcessing.ProcessStep(steps[i - stepAdjustment]);
-                }
-            }
-
-            return newSteps;
-        }
+    
 
     }
 }
